Move status bar color scheme to theme hook
